# KSregfastapi/src/openapi_server/models/data_delivery_mechanism.py
# coding: utf-8

"""
    Data registration service

    API for Data registration service. © 2024, O-RAN ALLIANCE. All rights reserved. 

    The version of the OpenAPI document: 2.0.0-alpha.1
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field
from typing import Any, ClassVar, Dict, List, Optional
from openapi_server.models.kafka_delivery_configuration import KafkaDeliveryConfiguration
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class DataDeliveryMechanism(BaseModel):
    """
    Defining the attributes of delivery mechanism supported
    """ # noqa: E501
    data_delivery_method: Optional[Any] = Field(description="Delivery Method supported", alias="dataDeliveryMethod")
    kafka_delivery_configuration: Optional[KafkaDeliveryConfiguration] = Field(default=None, alias="kafkaDeliveryConfiguration")
    __properties: ClassVar[List[str]] = ["dataDeliveryMethod", "kafkaDeliveryConfiguration"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of DataDeliveryMechanism from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of kafka_delivery_configuration
        if self.kafka_delivery_configuration:
            _dict['kafkaDeliveryConfiguration'] = self.kafka_delivery_configuration.to_dict()
        # set to None if data_delivery_method (nullable) is None
        # and model_fields_set contains the field
        if self.data_delivery_method is None and "data_delivery_method" in self.model_fields_set:
            _dict['dataDeliveryMethod'] = None

        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of DataDeliveryMechanism from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "dataDeliveryMethod": obj.get("dataDeliveryMethod"),
            "kafkaDeliveryConfiguration": KafkaDeliveryConfiguration.from_dict(obj.get("kafkaDeliveryConfiguration")) if obj.get("kafkaDeliveryConfiguration") is not None else None
        })
        return _obj